requirements.txt
beautifulsoup4==4.12.3
dashscope==1.13.6
faiss_cpu==1.7.4
html2text==2024.2.26
huggingface_hub==0.19.4
jieba==0.42.1
langchain==0.3.1
langchain_community==0.3.1
Markdown==3.6
modelscope==1.13.3
numpy==2.1.1
openai==1.50.2
Pillow==10.4.0
psutil==5.9.0
PyPDF2==3.0.1
python-dotenv==1.0.1
python_docx==1.1.0
python_pptx==1.0.0
Requests==2.32.3
sentence_transformers==3.0.1
streamlit==1.37.0
tiktoken==0.7.0
torch==2.1.2
tqdm==4.66.5
transformers==4.37.0
zhipuai==2.1.5.20230904
~==reamlit
~treamlit==1.30.0
~treamlit==1.36.0
